Moved files over into navigation-transitions folder to clean up components folder
